+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -34,6 +34,7 @@
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#MODIFIED_WARNING}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#USE_FOCUS_ACTIVATION}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#INCLUDE_ENTITY_VIEWER}
+ * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#INCLUDE_INSPECTOR}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#TRANSFER_FOCUS_ON_ENTER}
  * <li>{@link is.codion.swing.framework.ui.EntityPanel.Config#KEYBOARD_NAVIGATION}
  * <li>{@link is.codion.swing.framework.ui.EntityPanel.Config#DISPOSE_EDIT_DIALOG_ON_ESCAPE}
@@ -49,7 +50,7 @@
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#SUMMARY_PANEL_VISIBLE}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#INCLUDE_POPUP_MENU}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#INCLUDE_ENTITY_VIEWER}
- * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#INCLUDE_QUERY_INSPECTOR}
+ * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#INCLUDE_INSPECTOR}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#INCLUDE_CLEAR_CONTROL}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#INCLUDE_LIMIT_MENU}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#REFRESH_BUTTON_VISIBLE}